backend/fastapi-backend/requirements.txt
fastapi
uvicorn[standard]
python-jose
passlib[bcrypt]
psycopg2-binary
pydantic
sqlalchemy
alembic 